requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-filter==23.5
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
sendgrid==6.11.0
cloudinary==1.38.0
django-cloudinary-storage==0.3.0
Pillow==10.4.0
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0
coverage==7.3.2
